+++ CHANGELOG.md
@@ -34,8 +34,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Converter
   - .fc/.cf7 cfgyaml variants
 
+## [1.1.0]
+
 ### Added
 
+- GU14 GUIDs
 - Added `PriorityFeedbackBlocker` to IFO import
 - Added multiple `BuildBlocker` to IFO import
 - Support hitbox rotation for IFO import